refactor: BFF server責務を分割
